CMS-2685 Review: Let all get by key or id methods throw XxxxNotFoundException Removed unnecessary public modifiers

+++ modules/wem-core/src/main/java/com/enonic/wem/core/content/DeleteContentService.java
@@ -5,20 +5,26 @@ import javax.jcr.Session;
 
 import com.enonic.wem.api.command.content.DeleteContent;
 import com.enonic.wem.api.command.entity.DeleteNodeByPath;
-import com.enonic.wem.api.command.entity.DeleteNodeResult;
+import com.enonic.wem.api.command.entity.GetNodeByPath;
+import com.enonic.wem.api.content.Content;
+import com.enonic.wem.api.content.ContentNotFoundException;
 import com.enonic.wem.api.content.DeleteContentResult;
+import com.enonic.wem.api.content.UnableToDeleteContentException;
+import com.enonic.wem.api.entity.NoNodeAtPathFound;
+import com.enonic.wem.api.entity.Node;
 import com.enonic.wem.core.entity.DeleteNodeByPathService;
+import com.enonic.wem.core.entity.GetNodeByPathService;
 import com.enonic.wem.core.index.IndexService;
 
 
-class DeleteContentService
-extends ContentService
+public class DeleteContentService
+    extends ContentService
 {
     private final DeleteContent command;
 
     private final IndexService indexService;
 
-    DeleteContentService( final Session session, final DeleteContent command, final IndexService indexService )
+    public DeleteContentService( final Session session, final DeleteContent command, final IndexService indexService )
     {
         super( session );
         this.command = command;
@@ -30,17 +36,27 @@ extends ContentService
     {
         final DeleteNodeByPath deleteNodeByPathCommand =
             new DeleteNodeByPath( ContentNodeHelper.translateContentPathToNodePath( command.getContentPath() ) );
-        final DeleteNodeResult deleteNodeResult =
-            new DeleteNodeByPathService( this.session, this.indexService, deleteNodeByPathCommand ).execute();
 
-        switch ( deleteNodeResult )
+        try
         {
-            case NOT_FOUND:
-                return DeleteContentResult.NOT_FOUND;
-            case UNABLE_TO_DELETE:
-                return DeleteContentResult.UNABLE_TO_DELETE;
-            default:
-                return DeleteContentResult.SUCCESS;
+            final Node nodeToDelete =
+                new GetNodeByPathService( this.session, new GetNodeByPath( deleteNodeByPathCommand.getPath() ) ).execute();
+            final Content contentToDelete = CONTENT_TO_NODE_TRANSLATOR.fromNode( nodeToDelete );
+
+            if ( new ChildContentIdsResolver( this.session ).resolve( contentToDelete ).hasChildren() )
+            {
+                throw new UnableToDeleteContentException( command.getContentPath(), "Content has children" );
+            }
+
+            final Node deletedNode = new DeleteNodeByPathService( this.session, this.indexService, deleteNodeByPathCommand ).execute();
+
+            final Content deletedContent = CONTENT_TO_NODE_TRANSLATOR.fromNode( deletedNode );
+
+            return new DeleteContentResult( deletedContent );
+        }
+        catch ( NoNodeAtPathFound e )
+        {
+            throw new ContentNotFoundException( command.getContentPath() );
         }
     }
 }
